requirements(1).txt
streamlit==1.52.0
pyngrok==7.5.0
python-dotenv==1.2.1
langchain==0.2.16
langchain-core==0.2.41
langchain-community==0.2.11
langchain-text-splitters==0.2.4
langchain-ollama==0.1.1
requests==2.31.0
replicate==1.0.7
